Move WA handlers shared code to handlers/whatsapp.go
