+++ handlers/dialog360/dialog360.go
@@ -275,7 +275,7 @@ func (h *handler) Send(ctx context.Context, msg courier.Msg, clog *courier.Chann
 		msgParts = handlers.SplitMsgByChannel(msg.Channel(), msg.Text(), maxMsgLength)
 	}
 	qrs := msg.QuickReplies()
-	lang := utils.GetSupportedLanguage(msg.Locale())
+	lang := handlers.WAGetSupportedLanguage(msg.Locale())
 
 	var payloadAudio handlers.WACMTPayload
